renamed the sql folder
