+++ .github/workflows/continuous-integration.yml
@@ -12,7 +12,7 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Build Docker image
-        run: docker build -f ./Dockerfile.test -t docker-app .
+        run: docker build -f ./Dockerfile.test -t docker-backend-test .
 
       - name: Run Vitest inside Docker container
-        run: docker run docker-app
+        run: docker run docker-backend-test
