+++ src/database/createTables.ts
@@ -2,9 +2,9 @@
 import { Client } from "pg";
 
 // Project files
-import assignments from "../sql-queries/assignments";
-import candidates from "../sql-queries/candidates";
-import errorLogs from "../sql-queries/errorLogs";
+import assignments from "../sql/assignments";
+import candidates from "../sql/candidates";
+import errorLogs from "../sql/errorLogs";
 
 export default async function createTables(database: Client): Promise<void> {
   const message = "Postgres initialized tables";
