+++ src/sql/assignments.ts
