+++ src/sql/insertCandidate.ts
