+++ src/sql/insertErrorLog.ts
